Extract common interfaces or helper functions
